requirements.txt
fastapi
uvicorn
pytest
requests
gunicorn
coverage
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
python-dotenv